backend/requirements.txt
motor==3.3.1
pymongo==4.6.1
fastapi==0.115.4
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.1.0
stabilityai==0.1.0
